Moved DateTimeUtils to client side
